# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PRIVELUX API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: brand
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: visible
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/{id}/duplicate:
    post:
      operationId: duplicateProduct
      tags: [products]
      summary: Duplicate a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
  /products/{id}/related:
    get:
      operationId: listRelatedProducts
      tags: [products]
      summary: List related products by category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, price, categoryId, image, featured, visible, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        brandId:
          type: ["integer", "null"]
        brandName:
          type: ["string", "null"]
        image:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        featured:
          type: boolean
        visible:
          type: boolean
        stock:
          type: integer
        isOnSale:
          type: boolean
        originalPrice:
          type: ["number", "null"]
        salePrice:
          type: ["number", "null"]
        aiImageNotice:
          type: boolean
        hasSizes:
          type: boolean
        sizeTemplate:
          type: ["string", "null"]
        availableSizes:
          type: array
          items:
            type: string
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [name, price, categoryId, image]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        categoryId:
          type: integer
        brandId:
          type: ["integer", "null"]
        image:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        featured:
          type: boolean
        visible:
          type: boolean
        stock:
          type: integer
        isOnSale:
          type: boolean
        originalPrice:
          type: ["number", "null"]
        salePrice:
          type: ["number", "null"]
        aiImageNotice:
          type: boolean
        hasSizes:
          type: boolean
        sizeTemplate:
          type: ["string", "null"]
        availableSizes:
          type: array
          items:
            type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        categoryId:
          type: integer
        brandId:
          type: ["integer", "null"]
        image:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        featured:
          type: boolean
        visible:
          type: boolean
        stock:
          type: integer
        isOnSale:
          type: boolean
        originalPrice:
          type: ["number", "null"]
        salePrice:
          type: ["number", "null"]
        aiImageNotice:
          type: boolean
        hasSizes:
          type: boolean
        sizeTemplate:
          type: ["string", "null"]
        availableSizes:
          type: array
          items:
            type: string
...