# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PRIVELUX API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  brandId?: number | null;
  /** @nullable */
  brandName?: string | null;
  image: string;
  imageUrls?: string[];
  featured: boolean;
  visible: boolean;
  stock?: number;
  isOnSale?: boolean;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  salePrice?: number | null;
  aiImageNotice?: boolean;
  hasSizes?: boolean;
  /** @nullable */
  sizeTemplate?: string | null;
  availableSizes?: string[];
  createdAt: string;
}