# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PRIVELUX API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  description?: string;
  price: number;
  categoryId: number;
  /** @nullable */
  brandId?: number | null;
  image: string;
  imageUrls?: string[];
  featured?: boolean;
  visible?: boolean;
  stock?: number;
  isOnSale?: boolean;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  salePrice?: number | null;
  aiImageNotice?: boolean;
  hasSizes?: boolean;
  /** @nullable */
  sizeTemplate?: string | null;
  availableSizes?: string[];
}